[Modify] 토론 페이지를 메인 페이지로 대체
